+++ mopidy_podcast_itunes/backend.py
@@ -4,6 +4,7 @@ from mopidy import backend
 
 import pykka
 
+from .client import iTunesPodcastClient
 from .library import iTunesPodcastLibraryProvider
 
 
@@ -13,4 +14,5 @@ class iTunesPodcastBackend(pykka.ThreadingActor, backend.Backend):
 
     def __init__(self, config, audio):
         super(iTunesPodcastBackend, self).__init__()
-        self.library = iTunesPodcastLibraryProvider(config, backend=self)
+        self.client = iTunesPodcastClient(config)
+        self.library = iTunesPodcastLibraryProvider(config, self)