+++ CHANGES.rst
@@ -1,9 +1,14 @@
-v1.1.0 (UNRELEASED)
+v2.0.0 (UNRELEASED)
 -------------------
 
 - No longer use deprecated Mopidy-Podcast Directory API.
 
+- Add workaround for apparent changes in iTunes search API regarding
+  the ``explicit`` parameter.
+
 - Upgrade dependencies to Mopidy v1.1 and Mopidy-Podcast v2.0.
+
+- Add basic unit tests.
 
 
 v1.0.0 (2014-05-24)